OpenMAIC/anotherme2_engine/requirements.txt
# 依赖包要求
# LangChain 和相关
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# HTTP 请求
requests>=2.31.0

# Edge TTS
edge-tts>=6.1.0

# 音频处理
mutagen>=1.47.0

# 图结构构建（Scene Graph -> Geometry Graph）
networkx>=3.0

# 约束求解
sympy>=1.12

# Gateway (FastAPI + Postgres + Redis + S3/MinIO)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.2.0
redis>=5.0.0
boto3>=1.34.0
httpx>=0.27.0
python-multipart>=0.0.9
pillow>=10.0.0
pytest>=8.0.0
